Modify data.py and inference.py

+++ data.py
@@ -7,7 +7,7 @@ from datetime import datetime
 import pandas as pd
 import numpy as np
 import pickle
-import pyarrow as pa
+import os
 
 
 class ChildInstituteDataset(Dataset):
@@ -61,21 +61,25 @@ def scale(data, config: Dict[str, str], scaler: object = None) -> None:
     excluding_columnns = config.get('general').get('data').get('excluding_columns')
     target_columns = list(set(data.columns) - set(excluding_columnns))
 
-    if scaler is None:
-        scaler = getattr(preprocessing, config.get('train').get('data').get('scaler'))()
+    scaler_file_path = '/kaggle/input/models/acc,step=2,window_size=120,batch_size=256,lr=0.001,RADam/version_0/scaler.pickle'
+
+    if os.path.exists(scaler_file_path):
+        with open(scaler_file_path, 'rb') as file:
+            scaler = pickle.load(file)
+    else:
+        if scaler is None:
+            scaler = getattr(preprocessing, config.get('train').get('data').get('scaler'))()
+        with open(f'{config.get("general").get("log_dir")}/scaler.pickle', 'wb') as file:
+            pickle.dump(scaler, file)
 
     data.loc[:, target_columns] = scaler.fit_transform(data.filter(items=target_columns))
-
-    with open(f'{config.get("general").get("log_dir")}/scaler.pickle', 'wb') as file:
-        pickle.dump(scaler, file)
-
 
 
 def to_list(data, window_size: int, config: Dict[str, str], step: int = 1, key: List[str] = ['series_id'], scaler: object = None) -> List[pd.DataFrame]:
     data = [datum[1] for datum in data.groupby(key)]
     for datum in data:
         scale(datum, config, scaler)
-
+    
     slided_windows = []
     start_of_feature_index = np.where(data[0].columns.str.find('anglez') == 0)[0].item()
 
@@ -88,16 +92,25 @@ def to_list(data, window_size: int, config: Dict[str, str], step: int = 1, key: 
         in data]
     slided_windows.append(np.concatenate(slided_window))
 
+
     slided_window = [
             np.lib.stride_tricks.sliding_window_view(
-                datum.diff().iloc[1:, start_of_feature_index:-1],
+                datum[['anglez', 'enmo']].diff().iloc[1:],
                 window_size,
                 axis=0)[::step]
             for datum
             in data]
+    # slided_window = [
+    #         np.lib.stride_tricks.sliding_window_view(
+    #             datum.diff().iloc[1:, start_of_feature_index:-1],
+    #             window_size,
+    #             axis=0)[::step]
+    #         for datum
+    #         in data]
+    
     slided_windows.append(np.concatenate(slided_window))
     slided_windows = np.concatenate(slided_windows, axis=1, dtype=np.float32)
-
+    
     slided_windows_2 = []
     for win_size in range(window_size, 0, -30):
         slided_window = [
